create a hook to use OverviewTabSections from plugins
use const
