+++ frontend/packages/console-shared/src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './deep-compare-memoize';
 export * from './document-listener';
 export * from './fullscreen';
 export * from './scroll';
+export * from './plugins-overview-tab-section';
