+++ frontend/packages/dev-console/src/utils/resource-label-utils.ts
@@ -1,5 +1,6 @@
-import { K8sResourceKind } from '@console/internal/module/k8s';
 import * as _ from 'lodash';
+import { K8sResourceKind } from '@console/internal/module/k8s';
+import { TRIGGERS_ANNOTATION } from '@console/shared';
 
 export const getAppLabels = (
   name: string,
@@ -46,7 +47,7 @@ export const getCommonAnnotations = () => {
 };
 
 export const getTriggerAnnotation = (name: string) => ({
-  'image.openshift.io/triggers': JSON.stringify([
+  [TRIGGERS_ANNOTATION]: JSON.stringify([
     {
       from: { kind: 'ImageStreamTag', name: `${name}:latest` },
       fieldPath: `spec.template.spec.containers[?(@.name=="${name}")].image`,
