+++ frontend/packages/knative-plugin/src/components/overview/KnativeServiceResources.tsx
@@ -1,30 +1,14 @@
 import * as React from 'react';
 import * as _ from 'lodash';
 import { podPhase } from '@console/internal/module/k8s';
-import { isOverviewTabSection, useExtensions } from '@console/internal/plugins';
-import { AsyncComponent } from '@console/internal/components/utils';
 import { BuildOverview } from '@console/internal/components/overview/build-overview';
 import { PodModel } from '@console/internal/models';
-import { AllPodStatus, OverviewItem } from '@console/shared';
+import { AllPodStatus, OverviewItem, usePluginsOverviewTabSection } from '@console/shared';
 import { PodsOverview } from '@console/internal/components/overview/pods-overview';
 import RevisionsOverviewList from './RevisionsOverviewList';
 import KSRoutesOverviewList from './RoutesOverviewList';
 
 const REVISIONS_AUTOSCALED = 'All Revisions are autoscaled to 0';
-
-const getResourceTabSectionComp = (t) => (props) => (
-  <AsyncComponent {...props} loader={t.properties.loader} />
-);
-
-const usePluginTabSectionResource = (item: OverviewItem) => {
-  const tabSections = useExtensions(isOverviewTabSection);
-  return tabSections
-    .filter((section) => item[section.properties.key])
-    .map((section) => ({
-      Component: getResourceTabSectionComp(section),
-      key: section.properties.key,
-    }));
-};
 
 type KnativeServiceResourceProps = {
   item: OverviewItem;
@@ -40,7 +24,7 @@ const KnativeServiceResources: React.FC<KnativeServiceResourceProps> = ({ item }
   const linkUrl = `/search/ns/${namespace}?kind=${PodModel.kind}&q=${encodeURIComponent(
     `serving.knative.dev/${resKind.toLowerCase()}=${name}`,
   )}`;
-  const pluginComponents = usePluginTabSectionResource(item);
+  const pluginComponents = usePluginsOverviewTabSection(item);
   return (
     <>
       <PodsOverview
